+++ project/app/src/main/res/drawable/tab_selector.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+
+    <item android:drawable="@drawable/tab_indicator_selected"
+        android:state_selected="true"/>
+
+    <item android:drawable="@drawable/tab_indicator_default"/>
+</selector>
