+++ project/app/src/main/res/layout/user_onboarding_one.xml
@@ -2,7 +2,6 @@
 
 <ScrollView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="fill_parent"
     android:layout_height="fill_parent"
     android:background="#2E2E2E"
@@ -17,11 +16,11 @@
             android:id="@+id/textOnTop"
             android:layout_width="154dp"
             android:layout_height="77dp"
+            android:layout_marginTop="48dp"
             android:gravity="center"
-            android:text="A cada partida de tênis salve a sua pontuação"
+            android:text="@string/a_cada_partida_de_t_nis_salve_a_sua_pontua_o"
             android:textColor="#ffffff"
             android:textSize="15sp"
-            app:layout_constraintBottom_toTopOf="@+id/imageView2"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent" />
@@ -30,23 +29,25 @@
             android:id="@+id/textOnEnd"
             android:layout_width="184dp"
             android:layout_height="81dp"
+            android:layout_marginTop="28dp"
             android:gravity="center"
-            android:text="Mas você precisa estar registrado em algum clube para poder participar do Nexte!"
+            android:text="@string/mas_voc_precisa_estar_registrado_em_algum_clube_para_poder_participar_do_nexte"
             android:textColor="#ffffff"
             android:textSize="15sp"
-            app:layout_constraintBottom_toBottomOf="parent"
             app:layout_constraintEnd_toEndOf="parent"
+            app:layout_constraintHorizontal_bias="0.497"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toBottomOf="@id/imageView2" />
+            app:layout_constraintTop_toBottomOf="@+id/imageView2" />
 
         <ImageView
             android:id="@+id/imageView2"
-            android:layout_width="270dp"
-            android:layout_height="268dp"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintBottom_toBottomOf="parent"
+            android:layout_width="242dp"
+            android:layout_height="242dp"
+            android:layout_marginTop="28dp"
+            android:contentDescription="@string/imagem"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/textOnTop"
             app:srcCompat="@mipmap/ic_onboarding_preview" />
