+++ project/app/src/main/res/layout/user_onboarding_three.xml
@@ -20,9 +20,9 @@
             android:layout_height="77dp"
             android:layout_marginEnd="8dp"
             android:layout_marginStart="8dp"
-            android:layout_marginTop="32dp"
+            android:layout_marginTop="48dp"
             android:gravity="center"
-            android:text="Seu clube não está registrado? Entre em contato conosco!"
+            android:text="@string/seu_clube_n_o_est_registrado_entre_em_contato_conosco"
             android:textColor="#ffffff"
             android:textSize="15sp"
             app:layout_constraintEnd_toEndOf="parent"
@@ -34,24 +34,25 @@
             android:id="@+id/email_and_phone"
             android:layout_width="120dp"
             android:layout_height="34dp"
-            android:layout_marginBottom="8dp"
             android:layout_marginEnd="8dp"
             android:layout_marginStart="8dp"
+            android:layout_marginTop="10dp"
             android:gravity="center"
-            android:text="E-mail ou telefone"
+            android:text="@string/e_mail_ou_telefone"
             android:textColor="#c5c5c5"
-            app:layout_constraintBottom_toTopOf="@+id/phoneField"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent" />
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/textOnTop" />
 
         <View
             android:id="@+id/phone_line_left"
             android:layout_width="85dp"
             android:layout_height="1dp"
             android:layout_marginEnd="8dp"
+            android:layout_marginTop="17dp"
             android:background="#ffff"
             app:layout_constraintEnd_toStartOf="@+id/email_and_phone"
-            tools:layout_editor_absoluteY="167dp" />
+            app:layout_constraintTop_toTopOf="@+id/email_and_phone" />
 
         <View
             android:id="@+id/phone_line_right"
@@ -60,49 +61,57 @@
             android:layout_marginStart="8dp"
             android:background="#ffff"
             app:layout_constraintStart_toEndOf="@+id/email_and_phone"
-            tools:layout_editor_absoluteY="167dp" />
+            app:layout_constraintTop_toTopOf="@+id/phone_line_left" />
 
         <EditText
-            android:id="@+id/phoneField"
-            android:layout_width="317dp"
+            android:id="@+id/contactField"
+            android:layout_width="300dp"
             android:layout_height="27dp"
-            android:layout_marginEnd="60dp"
-            android:layout_marginStart="60dp"
-            android:layout_marginTop="188dp"
+            android:layout_marginTop="10dp"
             android:background="@drawable/edge_onboard"
             android:ems="10"
-            android:inputType="textPassword"
+            android:inputType="text"
+            android:labelFor="@+id/email_and_phone"
+            android:textColor="#FFF"
+            android:textColorHint="#DDD"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
+            app:layout_constraintTop_toBottomOf="@+id/email_and_phone"
+            tools:ignore="LabelFor" />
 
         <EditText
             android:id="@+id/requestField"
-            android:layout_width="317dp"
+            android:layout_width="300dp"
             android:layout_height="155dp"
             android:layout_marginEnd="60dp"
             android:layout_marginStart="60dp"
-            android:layout_marginTop="280dp"
+            android:layout_marginTop="10dp"
             android:background="@drawable/edge_onboard"
             android:ems="10"
-            android:inputType="textPassword"
+            android:gravity="top"
+            android:inputType="textMultiLine"
+            android:labelFor="@+id/requestText"
+            android:singleLine="false"
+            android:textColor="#fff"
+            android:textColorHint="#ddd"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
+            app:layout_constraintTop_toBottomOf="@+id/requestText"
+            tools:ignore="LabelFor" />
 
         <TextView
             android:id="@+id/requestText"
             android:layout_width="151dp"
-            android:layout_height="40dp"
-            android:layout_marginBottom="8dp"
+            android:layout_height="34dp"
             android:layout_marginEnd="8dp"
             android:layout_marginStart="8dp"
+            android:layout_marginTop="24dp"
             android:gravity="center"
-            android:text="Solicitação de registro"
+            android:text="@string/solicita_o_de_registro"
             android:textColor="#c5c5c5"
-            app:layout_constraintBottom_toTopOf="@+id/requestField"
             app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintStart_toStartOf="parent" />
+            app:layout_constraintStart_toStartOf="parent"
+            app:layout_constraintTop_toBottomOf="@+id/contactField" />
 
         <View
             android:id="@+id/request_line_left"
@@ -111,30 +120,31 @@
             android:layout_marginStart="8dp"
             android:background="#ffff"
             app:layout_constraintStart_toEndOf="@+id/requestText"
-            tools:layout_editor_absoluteY="255dp" />
+            app:layout_constraintTop_toTopOf="@+id/request_line_right" />
 
         <View
             android:id="@+id/request_line_right"
             android:layout_width="75dp"
             android:layout_height="1dp"
             android:layout_marginEnd="8dp"
+            android:layout_marginTop="17dp"
             android:background="#ffff"
             app:layout_constraintEnd_toStartOf="@+id/requestText"
-            tools:layout_editor_absoluteY="255dp" />
+            app:layout_constraintTop_toTopOf="@+id/requestText" />
 
         <Button
             android:id="@+id/sentRequestButton"
-            android:layout_width="317dp"
-            android:layout_height="21dp"
+            android:layout_width="300dp"
+            android:layout_height="32dp"
             android:layout_marginEnd="60dp"
             android:layout_marginStart="60dp"
-            android:layout_marginTop="472dp"
+            android:layout_marginTop="24dp"
             android:background="@drawable/edge_onboard"
-            android:text="Enviar"
+            android:text="@string/enviar"
             android:textColor="#ffff"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent" />
+            app:layout_constraintTop_toBottomOf="@+id/requestField" />
 
 
     </android.support.constraint.ConstraintLayout>
